chore: finish security logic, add auth controller

+++ src/main/kotlin/com/cloudrader/inventarium/api/v1/NoteController.kt
@@ -3,6 +3,7 @@ package com.cloudrader.inventarium.api.v1
 import com.cloudrader.inventarium.model.Note
 import com.cloudrader.inventarium.repository.NoteRepository
 import org.bson.types.ObjectId
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -39,6 +40,7 @@ class NoteController(
     fun create(
         @RequestBody body: NoteRequest
     ): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication!!.principal as String
         val note = repository.save(
             Note(
                 id = body.id?.let { ObjectId(it) } ?: ObjectId.get(),
@@ -46,7 +48,7 @@ class NoteController(
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId()
+                ownerId = ObjectId(ownerId)
             )
         )
 
@@ -54,9 +56,8 @@ class NoteController(
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication!!.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -64,7 +65,14 @@ class NoteController(
 
     @DeleteMapping(path = ["/{id}"])
     fun delete(@PathVariable id: String) {
-        repository.deleteById(ObjectId(id) )
+        val note = repository.findById(ObjectId(id)).orElseThrow {
+            IllegalArgumentException("Note not found")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication!!.principal as String
+
+        if(note.ownerId.toHexString() == ownerId) {
+            repository.deleteById(ObjectId(id))
+        }
     }
 
     private fun Note.toResponse() : NoteController.NoteResponse {
